make configurationError more prominent, create a base impl
